Salt remains the same across sessions
For Fernet to work correctly, the salt must be the same when
re-encrypting the project after a successful decryption.

This is because the key for decryption and encryption must be the
same during a single session. So if the project was decrypted with
one salt, the we want to generate a new salt to re-encrypt the file
with a new key, Fernet will cough up an exception. Presumably this
problem won't exist with other crypto-systems (e.g. AES-GCM).